+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Version History
        improved GPU buffer support
     -  `oidnExecuteFilterAsync` and `oidnSyncDevice` for asynchronous
        filter execution
+-   Added `type` device parameter for querying the type of the created device 
 -   Fixed over-conservative buffer bounds checking for images with custom strides
 
 ### Changes in v1.4.3: